==> 10-Summary.ipynb <==
import os
import pandas as pd
def mean(a):
    return sum(a) / len(a)

#timeseries baseline
results_dir = "results/timeseries-baseline"
files = os.listdir(results_dir)
tasks_list = []
for file in files:
    if not file.endswith('.p'):
        continue
    task_result = pd.read_pickle(os.path.join(results_dir, file))        
    task_result["hidden_units"], task_result["model_type"], task_result["prediction_task"], task_result["iteration"] = file.split("-")[:4]
    task_result["model_category"] = "baseline"
    tasks_list.append(task_result)
results_df = pd.DataFrame(tasks_list)    
functions = ['mean','std']
baseline_results = results_df.groupby(['model_type','model_category','prediction_task'], as_index = False).agg({"auc":functions, "auprc":functions, "F1":functions, 'acc':functions})
baseline_results.columns = baseline_results.columns.map('_'.join).str.strip('_')
column_list = ['auc_mean', 'auc_std', 'auprc_mean', 'auprc_std','F1_mean', 'F1_std', 'acc_mean', 'acc_std']
for column in column_list:
       baseline_results[column] = baseline_results[column] * 100

prediction_task_rename = {'los_3': 'LOS > 3 Days', "los_7": "LOS > 7 Days","mort_hosp": "In-Hospital Mortality","mort_icu": "In-ICU Mortality"}
baseline_results['prediction_task'] = baseline_results['prediction_task'].map(prediction_task_rename)
print(baseline_results)

results_dir = "results/multimodal-baseline"
files = os.listdir(results_dir)
tasks_list = []
for file in files:
    if not file.endswith('.p'):
        continue
    task_result = pd.read_pickle(os.path.join(results_dir, file))        
    task_result["layer"], task_result["hidden_units"], task_result["embedding"], task_result["prediction_task"], task_result["iteration"]  = file.split("-")[:5]
    task_result["model_category"] = "MultiModal Baseline"
    tasks_list.append(task_result)
results_df = pd.DataFrame(tasks_list) 
functions = ['mean','std']
mmbaseline = results_df.groupby(['embedding','model_category','prediction_task'], as_index = False).agg({"auc":functions, "auprc":functions, "F1":functions, 'acc':functions})
mmbaseline.columns = mmbaseline.columns.map('_'.join).str.strip('_')
column_list = ['auc_mean', 'auc_std', 'auprc_mean', 'auprc_std','F1_mean', 'F1_std', 'acc_mean', 'acc_std']
for column in column_list:
       mmbaseline[column] = mmbaseline[column] * 100
prediction_task_rename = {'los_3': 'LOS > 3 Days', "los_7": "LOS > 7 Days","mort_hosp": "In-Hospital Mortality","mort_icu": "In-ICU Mortality"}
mmbaseline['prediction_task'] = mmbaseline['prediction_task'].map(prediction_task_rename)
print(mmbaseline)

#proposed model
results_dir = "results/cnn"
files = os.listdir(results_dir)
tasks_list = []
for file in files:
    if not file.endswith('.p'):
        continue
    task_result = pd.read_pickle(os.path.join(results_dir, file))        
    task_result["model_type"], task_result["hidden_units"], task_result["embedding"], task_result["prediction_task"], task_result["iteration"], task_result["ner_type"]  = file.split("-")[:6]
    task_result["model_category"] = "Proposed Model"
    tasks_list.append(task_result)
results_df = pd.DataFrame(tasks_list) 
functions = ['mean','std']
proposed_model = results_df.groupby(['embedding','model_category','prediction_task'], as_index = False).agg({"auc":functions, "auprc":functions, "F1":functions, 'acc':functions})
proposed_model.columns = proposed_model.columns.map('_'.join).str.strip('_')
column_list = ['auc_mean', 'auc_std', 'auprc_mean', 'auprc_std','F1_mean', 'F1_std', 'acc_mean', 'acc_std']
for column in column_list:
       proposed_model[column] = proposed_model[column] * 100
prediction_task_rename = {'los_3': 'LOS > 3 Days', "los_7": "LOS > 7 Days","mort_hosp": "In-Hospital Mortality","mort_icu": "In-ICU Mortality"}
proposed_model['prediction_task'] = proposed_model['prediction_task'].map(prediction_task_rename)
print(proposed_model)
